python/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
numpy>=1.24.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
git+https://github.com/ChaoningZhang/MobileSAM.git
